Removing linker class (using higher-order functions intead). Renaming the visitor to use more standard names (enter(), exit())
